Adding this for comparison to prev 2020-2022 list
When aggregating over more years the number of suicide per county increases and so several counties that were previously below the threshold for statistical reliability to be estimated per 100,000 are now added to the list. Once this adjustment is made the four counties with highest suicide rate are all counties with a reservation.